Remove part of unnecessary transpositions, computations and memory allocations from batch SVD (#2055)
There were a lot of unnecessary matrix transpositions and memory allocations in current SVD implementation. Some cases also include extra computations.

This first step removes transpositions from compute_seq function that is used when nRows < 2 * nColumns in the input matrix in batch SVD.

SVD(A) = U * S * V'T, where U and V are orthogonal matrices, S is a diagonal matrix and V'T is a transposed matrix V.
SVD(A'T) = V * S * U'T - this knowledge is used in the changes to remove extra transpositions.

Also, less computations and memory allocations are performed if not all the results of the decomposition are requested by the user. For example, when PCA is computed, left singular vectors (U) are not needed.