Support building reference backend with clang (#2529)
* Add correct compile flags for clang reference build

The clang makefile doesn't have the right includes defined for the
reference backend implementation. Update this to match the defines for
GCC.

* Use right include directory for TBB

The makefile doesn't use the most up-to-date variable for finding TBB
libraries. This meant that a user-installed TBB, specified through
TBBROOT, was not picked up during the build process. Now, if there is no
TBB in the __deps folder, it is picked up from TBBROOT.