Move makefiles into x86_64 target directory
It would be nice to be able to build oneDAL on different CPU targets. To
make it easier to develop build files for different targets, we move the
existing build files into dev/make/x86_64.

The top level makefile takes a TARGET parameter, which at the moment can
only be x86_64, and defaults to this. This means that the build commands
do not change with this. Any future changes to the builds need to be
done in the dev/make/x86_64 directory.