Fix csr_accessor crash when moving data from USM or host memory to host memory (#2662)
The implementation of copy_all2all backend function was changed: copy(...) was replaces with memcpy(...) that copies exactly the requested number of bytes. In the previous implementation the copy(...) function was multiplying the requested number of bytes to the type size which produces the wrong size of the copied array.

csr_accessor test was extended to cover the cases that were failing previously.