Makefile refactoring to factor out common build code (#2672)
* Don't use double brackets in source script

Match the same style as used in the rest of the script. I observed a
syntax error when running in a Docker container locally as well, so fix
this up quickly.

* Move arch and uarch specific code out of main Makefile

Following on from pull request
https://github.com/oneapi-src/oneDAL/pull/2614, we want to make it
easier to distinguish common and platform specific code when building.
We take common code from the main makefile, and split this out into arch
and uarch specific files. We try and share as much code as possible
between uarch files.

In this first commit, we make sure that the x86 build works on linux.
Next commits will make sure that the Mac, Windows and Arm builds also
work. For now these are broken.

* Update Mac, Windows and Arm makefile configuration

Put Mac and Windows makefile code into their own files, and correct some
copy-paste errors in the Arm build. This should build on all platforms
now, but there are some more commits to go on top before the refactor is
done.

* Refactor compiler configuration in makefiles

We want to have as much common code shared between makefiles for
different compilers, as well as for the different backends. We shuffle
things around a little. This now passes on x86 Linux as well as aarch64.
Needs some internal tests to be run.

* Remove recursive definition

* Code review copyright notice and new line changes

* Remove year from contributors copyright notice

* Don't use year in copyright

* review comments