Only load conda environment if required
The build script for CI is useful to use locally. However, the conda
path does not need to be the same on the local system. As such, we only
source the conda environment in the case that we request it.

Locally, having conda available in PATH will prevent conda from being
sourced again.