Set the compilers early in CI build script
The combination of architecture, compiler family and whether we are
performing cross-compilation allows us to set the compiler in the
top-level CI build script early on. This means that the same toolchain
is used to build all of the components that need building. This is
mainly of use in the case that we are using a reference (i.e. OpenBLAS)
backend on a non-x86 system, in which case OpenBLAS and oneTBB are built
from source.