Allow test.sh CI script to run locally
So that it is easier to run the CI test scripts on a developer machine,
a few changes are made to the script. As well as making some functional
changes, shellcheck is run and some escaping in the script is cleaned
up.

Running the script locally makes it easier to check that the tests are
running as expected when porting to other architectures, before pushing
changes to a pull request.

The main changes are:

* Made paths relative to the oneDAL root directory, to allow an
  out-of-source run
* Only source miniconda in the case that the file exists. This means
  that the CI runs are the same, but that a developer doesn't need to
  install miniconda on their own system
* Write stderr from tests to file. The redirects were in the wrong order
  for this to happen

Minor changes are:

* Added help text
* Iterate over a list of elements, rather than a string
* Update escaping, as suggested by shellcheck
* Add some exit paths (e.g. if directories do not exist)