Add Arm LLVM cross-compilation CI pipeline
For Clang cross-compilation we need to do a few extra things when
compared to GNU. First of all, we need to have a sysroot with all the
libraries installed that we want. As such, we debootstrap a sysroot,
which is to be cached in the Azure CI pipelines. We also install a more
recent version of the LLVM compiler. Whilst not necessary for the Arm
build, this is necessary to get some more recent support for RISC-V, in
particular the vector extensions, which will be coming in a future
commit.

Changes made:
* Install clang from llvm.org, rather than default apt
* Build a sysroot with necessary packages, and cache this in the CI
* Add a CMake toolchain file for Clang on aarch64
* Update OpenBLAS build to pass appropriate options for clang
  cross-compilation
* Add a variable to the oneDAL makefile to detect Clang
  cross-compilation, and set appropriate flags for building
* Update list of failing tests under emulation