Refactor OpenBLAS build script (#2711)
* Refactor OpenBLAS build script

Getting ready to add a clang cross-compile path, we tidy up the OpenBLAS
build script a little. As bonus, because I recently learned about
shellcheck, run this on the openblas.sh and build.sh scripts and tidy up
warnings. The shellcheck changes are mostly to do with escaping.

A summary of the openblas changes are:

* Add help text. Use the 'column' command to indent this properly
* Make the command-line arguments use a '-' instead of '_' to match
  build.sh
* Add an option to set the source directory of OpenBLAS. This makes it
  easier to run this script locally, and build out-of-source
* Make paths relative to the oneDAL root directory, rather than the
  current working directory. This enables out-of-source builds
* Change the install directory to at least contain the architecture
  being built for. Also update the build script to point to the correct
  custom install location. This just makes things a little neater when
  building for different platforms.

* Use a default target architecture in OpenBLAS build

The architecture can be read from `uname` in the case that none is
supplied. This is required to satisfy the calling of the build without
any parameters provided, in the Docker development environment.

* Use same make options in clean

* Use most recent link to OpenBLAS repo

* Pass prefix and version as optional parameters