Update doc/readme.md to define how to create cross references (#2729)
Instructions on how to define an explicit target for cross references so that duplicate targets are not created.

Using implicit targets (using the heading as the link you put in your cross reference) introduces the possibility of having two targets with the same name. For example, if you have a section titled "Configuration" in the introduction section, but you have another "Configuration" section in the API section, then the reference will break because the generator (Sphinx) will not know which target to point to. Explicit targets solves this problem.

Some lines you may not see a change because the only change is that a trailing space at the end of the line has been removed. This is the default in my text editor because I work with some repositories that require that no file has a trailing space.