CI fixes in OpenBLAS builds
The riscv64 and x86_64 OpenBLAS builds have been failing. For the riscv64
build, the location of the qemu debian package changed. We update the
URL to obtain, as well as making sure that failing to obtain the right
QEMU will cause an error in CI.

For the x86_64 build, illegal instructions are being reported, which I
can't reproduce locally. My intuition is that there is something wrong
with the cached build, so we update the key for the build. The naming in
the key now matches that of the aarch64 and riscv64 builds.