Changed openblas build to ILP64 (#2801)
* Changed openblas build to ILP64

Default openblas build is LP64 on 64 bit machines. But DAAL_INT is 64-bit
signed integer. This causes errors during openblas calls (mainly LAPACK calls),
where the argument is expected to be 32-bit integers or pointers to 32-bit integers,
but a 64-bit integer is given. This causes many examples that use LAPACK routines
to fail (when using openblas as the BLAS/LAPACK backend).

An option to switch ILP64 interface has been added to the openblas build script.

The examples that are passing have also been removed from the exclude list.

Signed-off-by: Dhanus M Lal <Dhanus.MLal@fujitsu.com>

* added ILP64 to the cache key for gcc+openblas builds

Signed-off-by: Dhanus M Lal <Dhanus.MLal@fujitsu.com>

* updated cache key and changed qemu version

Changed qemu package version and added the ILP64 tag to openblas cache key
for the LLVM+openblas builds

Signed-off-by: Dhanus M Lal <Dhanus.MLal@fujitsu.com>

---------

Signed-off-by: Dhanus M Lal <Dhanus.MLal@fujitsu.com>